refactor: rename service_provider_tests.py to helpers.py
Clearer name for the shared test mixins and helpers module used
across Matter, Schlage, and Akuvox provider tests.
